refactor: the friends-request and friends features
