+++ src/features/friends/friends.controller.ts
@@ -1,7 +1,6 @@
 import { Controller, Get, Param, Post, Request } from '@nestjs/common';
 import { FriendsService } from './friends.service';
 import { Friends } from './models/friend.model';
-import { FriendsRequest } from './models/friend-request.model';
 
 @Controller('friends')
 export class FriendsController {
@@ -10,16 +9,5 @@ export class FriendsController {
   @Get()
   findAll(): Promise<Friends[]> {
     return this.friendsService.findAll();
-  }
-
-  @Post('sendAddFriendRequest/:userEmailToAdd')
-  addFriend(
-    @Param('userEmailToAdd') userEmailToAdd,
-    @Request() req,
-  ): Promise<FriendsRequest> {
-    return this.friendsService.sendAddFriendRequest(
-      req.user.sub,
-      userEmailToAdd,
-    );
   }
 }
