+++ src/features/friends/friends.module.ts
@@ -5,7 +5,7 @@ import { FriendsService } from './friends.service';
 import { FriendsController } from './friends.controller';
 import { UsersService } from '../users/users.service';
 import { User } from '../users/user.model';
-import { FriendsRequest } from './models/friend-request.model';
+import { FriendsRequest } from '../friends-request/model/friend-request.model';
 
 @Module({
   imports: [SequelizeModule.forFeature([Friends, User, FriendsRequest])],
